requirements.txt
appdirs==1.4.4
attrs==20.3.0
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
defusedxml==0.7.1
idna==2.10
isodate==0.6.0
lxml==4.6.2
polarion==0.1.8
progressbar2==3.53.1
python-docx==0.8.10
python-utils==2.5.6
pytz==2021.1
requests==2.25.1
requests-file==1.5.1
requests-toolbelt==0.9.1
six==1.15.0
urllib3==1.26.3
zeep==4.0.0
